Rename to more proper name, add .gitignore
